+++ TTS/src/utils/inference_utils.py
@@ -53,9 +53,9 @@ class HomophonesReplacer:
 
     Creation process of homophones_map.json:
 
-    1. Establish a word corpus using the [Tencent AI Lab Embedding Corpora v0.2.0 large] with 12 million entries. After cleaning, approximately 1.8 million entries remain. Use ChatTTS to infer the text.
+    1. Establish a word corpus using the [Tencent AI Lab Embedding Corpora v0.2.0 large] with 12 million entries. After cleaning, approximately 1.8 million entries remain.
     2. Record discrepancies between the inferred and input text, identifying about 180,000 misread words.
-    3. Create a pinyin to common characters mapping using correctly read characters by ChatTTS.
+    3. Create a pinyin to common characters mapping using correctly read characters by LinguifySpeech.
     4. For each discrepancy, extract the correct pinyin using [python-pinyin] and find homophones with the correct pronunciation from the mapping.
 
     Thanks to:
@@ -73,12 +73,16 @@ class HomophonesReplacer:
 
     def replace(self, text):
         result = []
+        replaced_words = []
+
         for char in text:
             if char in self.homophones_map:
-                result.append(self.homophones_map[char])
+                 repl_char = self.homophones_map[char]
+                 result.append(repl_char)
+                 replaced_words.append((char, repl_char))
             else:
                 result.append(char)
-        return ''.join(result)
+        return ''.join(result), replaced_words
 
 
 def count_invalid_characters(s):